Extract RawContactQueries
